Actual distance has its own component, cannot click 0-mile past dates
